Base Schema 추가 및 AuthMixin 추가
