+++ schemas/post.py
@@ -1,24 +1,10 @@
-from uuid import uuid4, UUID
+from pydantic import BaseModel, Field, field_validator
+from datetime import datetime
 
-from pydantic import BaseModel, Field, validator, field_validator
-from datetime import datetime, timezone
-
-
-def _default_time():
-    return datetime.now(timezone.utc)
+from schemas.base import BaseDataInfo, AuthMixin
 
 
-def _change_local_time(time: datetime):
-    return time.astimezone()
-
-
-class PostBase(BaseModel):
-    id: UUID = Field(default_factory=uuid4)
-    created_at: datetime = Field(default_factory=_default_time)
-    updated_at: datetime = Field(default_factory=_default_time)
-
-
-class Post(PostBase):
+class Post(BaseDataInfo, AuthMixin):
     author: str = Field(min_length=1, max_length=20)
     title: str = Field(min_length=5, max_length=50)
     content: str = Field(min_length=1)
